Move incremental SMT ast into sub directory
So that the ast data structure files are kept together and separated
from the non data structure components of the incremental SMT decision
procedure.
