+++ src/solvers/smt2_incremental/ast/smt_index.cpp
