+++ src/solvers/smt2_incremental/ast/smt_logics.cpp
@@ -1,11 +1,12 @@
 // Author: Diffblue Ltd.
 
-#include <solvers/smt2_incremental/smt_logics.h>
+#include "smt_logics.h"
 
 // Define the irep_idts for logics.
 #define LOGIC_ID(the_id, the_name)                                             \
   const irep_idt ID_smt_logic_##the_id{"smt_logic_" #the_id};
-#include <solvers/smt2_incremental/smt_logics.def>
+#include "smt_logics.def"
+
 #undef LOGIC_ID
 
 bool smt_logict::operator==(const smt_logict &other) const
@@ -26,7 +27,7 @@ void accept(const smt_logict &logic, const irep_idt &id, visitort &&visitor)
     return visitor.visit(static_cast<const smt_logic_##the_id##t &>(logic));
 // The include below is marked as nolint because including the same file
 // multiple times is required as part of the x macro pattern.
-#include <solvers/smt2_incremental/smt_logics.def> // NOLINT(build/include)
+#include "smt_logics.def" // NOLINT(build/include)
 #undef LOGIC_ID
   UNREACHABLE;
 }
