+++ src/solvers/smt2_incremental/ast/smt_logics.def
