+++ src/solvers/smt2_incremental/ast/smt_terms.def
