+++ src/solvers/smt2_incremental/smt2_incremental_decision_procedure.h
@@ -9,9 +9,9 @@
 #include <util/message.h>
 #include <util/std_expr.h>
 
+#include <solvers/smt2_incremental/ast/smt_terms.h>
 #include <solvers/smt2_incremental/object_tracking.h>
 #include <solvers/smt2_incremental/smt_object_size.h>
-#include <solvers/smt2_incremental/smt_terms.h>
 #include <solvers/smt2_incremental/type_size_mapping.h>
 #include <solvers/stack_decision_procedure.h>
