+++ src/solvers/smt2_incremental/smt_response_validation.h
@@ -5,8 +5,8 @@
 
 #include <util/nodiscard.h>
 
+#include <solvers/smt2_incremental/ast/smt_responses.h>
 #include <solvers/smt2_incremental/response_or_error.h>
-#include <solvers/smt2_incremental/smt_responses.h>
 
 NODISCARD response_or_errort<smt_responset> validate_smt_response(
   const irept &parse_tree,
