+++ unit/solvers/smt2_incremental/ast/module_dependencies.txt
@@ -1,0 +1,2 @@
+solvers/smt2_incremental/ast
+testing-utils
