+++ unit/solvers/smt2_incremental/ast/smt_terms.cpp
@@ -2,8 +2,8 @@
 
 #include <util/mp_arith.h>
 
+#include <solvers/smt2_incremental/ast/smt_terms.h>
 #include <solvers/smt2_incremental/smt_core_theory.h>
-#include <solvers/smt2_incremental/smt_terms.h>
 #include <testing-utils/invariant.h>
 #include <testing-utils/use_catch.h>
